+++ lib/ts/types.ts
@@ -13,6 +13,7 @@
  * under the License.
  */
 
+import type { DateProviderInput } from "./dateProvider/types";
 import type { BaseRecipeModule } from "./recipe/recipeModule/baseRecipeModule";
 import type { NormalisedConfig as NormalisedRecipeModuleConfig } from "./recipe/recipeModule/types";
 import type { TranslationFunc, TranslationStore } from "./translation/translationHelpers";
@@ -89,6 +90,8 @@ export type SuperTokensConfig = {
     cookieHandler?: CookieHandlerInput;
 
     windowHandler?: WindowHandlerInput;
+
+    dateProvider?: DateProviderInput;
 
     usesDynamicLoginMethods?: boolean;