+++ test/with-typescript/src/App.tsx
@@ -59,6 +59,8 @@ import {
 import EmailVerification from "../../../recipe/emailverification";
 import MultiFactorAuth from "../../../recipe/multifactorauth";
 
+import { DateProviderReference } from "../../../utils/dateProvider";
+import { DateProviderInput, DateProviderInterface } from "../../../utils/dateProvider/types";
 /*
  * This application is used with the purpose of illustrating Supertokens with typescript.
  * It is also used internally for deploy previews, hence a lot of code you will see
@@ -99,6 +101,16 @@ const rid = window.localStorage.getItem("rid") || "emailpassword";
 
 const recipeList = getRecipeList();
 
+const dateProviderImplementation: DateProviderInterface = {
+    getThresholdInSeconds: () => 0,
+    setThresholdInSeconds: () => {},
+    getClientClockSkewInMillis: () => 0,
+    setClientClockSkewInMillis: () => {},
+    now: () => Date.now(),
+};
+
+const dateProviderInput: DateProviderInput = () => dateProviderImplementation;
+
 SuperTokens.init({
     appInfo: {
         appName: "SuperTokens Demo App",
@@ -112,6 +124,7 @@ SuperTokens.init({
         return undefined;
     },
     recipeList,
+    dateProvider: dateProviderInput,
 });
 
 function App() {
@@ -340,8 +353,8 @@ function getRecipeList() {
                         shouldDoInterceptionBasedOnUrl: (...input) => {
                             return oI.shouldDoInterceptionBasedOnUrl(...input);
                         },
-                        calculateClockSkewInMillis: (input) => {
-                            return oI.calculateClockSkewInMillis(input);
+                        calculateClockSkewInMillis: (...input) => {
+                            return oI.calculateClockSkewInMillis(...input);
                         },
                     };
                 },
@@ -527,7 +540,9 @@ function getThirdPartyConfigs() {
                 ThirdParty.Github.init(),
                 ThirdParty.Google.init({
                     id: "some client ID",
-                    buttonComponent: ({ name }) => <span>ASDF {name}</span>,
+                    buttonComponent: ({ name }) => {
+                        return <div>ASDF {name}</div>;
+                    },
                 }),
                 ThirdParty.Facebook.init(),
                 ThirdParty.Apple.init(),